Fixed All Issues and Added Tests

+++ apps/api/src/main/java/com/agricloud/controller/PlotController.java
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.*;
 
 import com.agricloud.model.PlotModel;
 import com.agricloud.response.GeneralResponse;
+import com.agricloud.service.AccountService;
 import com.agricloud.service.PlotService;
 
 @RestController
@@ -13,34 +14,61 @@ public class PlotController {
 
     @Autowired
     private PlotService plotService;
+
+    @Autowired
+    private AccountService accountService;
     
     @GetMapping("/{name}/info")
     public GeneralResponse info (
         @PathVariable(value = "name") String plotName,
-        @RequestParam(value = "accID") int accID
+        @RequestParam(value = "accID") String accUID
     ) {
+        Integer accID = accountService.accountID(accUID);
+        if (accID == null) {
+            return new GeneralResponse("Access forbidden!", null); 
+        }
+
         return plotService.info(plotName, accID);
     }
 
     @GetMapping("/{name}/status")
     public GeneralResponse logs (
-        @PathVariable(value = "name") int plotID, 
+        @PathVariable(value = "name") String plotName, 
         @RequestParam(value = "hours", defaultValue = "1") int pastHours,
-        @RequestParam(value = "accID") int accID
+        @RequestParam(value = "accID") String accUID
     ) {
-        return plotService.status(plotID, pastHours, accID);
+        Integer accID = accountService.accountID(accUID);
+        if (accID == null) {
+            return new GeneralResponse("Access forbidden!", null); 
+        }
+
+        return plotService.status(plotName, pastHours, accID);
     }
     
     @PostMapping("/create")
-    public GeneralResponse create (@RequestBody PlotModel plot) {
+    public GeneralResponse create (
+        @RequestBody PlotModel plot,
+        @RequestParam(value = "accID") String accUID
+    ) {
+        Integer accID = accountService.accountID(accUID);
+        if (accID == null) {
+            return new GeneralResponse("Access forbidden!", null); 
+        }
+
+        plot.setAccountID(accID);
         return plotService.create(plot);
     }
 
     @PostMapping("/{name}/terminate") 
     public GeneralResponse terminate (
         @PathVariable(value = "name") String plotName,
-        @PathVariable(value = "accID") int accID
+        @RequestParam(value = "accID") String accUID
     ) {
+        Integer accID = accountService.accountID(accUID);
+        if (accID == null) {
+            return new GeneralResponse("Access forbidden!", null); 
+        }
+
         return plotService.terminate(plotName, accID);
     }
